requirements.txt
# Core Dependencies
python-telegram-bot==20.7
python-binance==1.0.19
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Database & ORM
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Data Science & ML
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0

# Technical Analysis
ta-lib==0.4.28
talib-binary==0.4.19

# Async & Scheduling
schedule==1.2.0
asyncio==3.4.3

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging & Monitoring
structlog==23.2.0
colorlog==6.8.0

# Security & Encryption
cryptography==41.0.8
bcrypt==4.1.2

# HTTP & API
requests==2.31.0
urllib3==2.1.0

# Data Validation & Serialization
marshmallow==3.20.2
jsonschema==4.20.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0

# Performance & Optimization
uvloop==0.19.0
cython==3.0.6

# Additional ML Libraries
xgboost==2.0.2
lightgbm==4.1.0

# Visualization (Optional)
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.13.0

# Web Framework (for future dashboard)
flask==3.0.0
flask-cors==4.0.0

# Memory & Performance
psutil==5.9.6
memory-profiler==0.61.0

# Time Series Analysis
statsmodels==0.14.1
arch==6.2.0

# Additional Utilities
tqdm==4.66.1
rich==13.7.0
typer==0.9.0
